+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added a custom difficulty to allow you to configure all the difficulty settings
 - Updated some personality names to prevent word wrapping with new fonts
+- Galaxy size no longer has any effect on the chance of a FFA occuring
 
 ## v1.3.0 - 2019-10-11