+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Added a custom difficulty to allow you to configure all the difficulty settings
 - Updated some personality names to prevent word wrapping with new fonts
 - Galaxy size no longer has any effect on the chance of a FFA occuring
+- Fixed the enemy system owner sometimes spawning with multiple Commanders
+  - Note that this breaks compatibility with all wars started pre-v1.0
 
 ## v1.3.0 - 2019-10-11